Improve image-loader controller structure
