+++ image-loader/app/controllers/ImageLoaderController.scala
@@ -1,31 +1,28 @@
 package controllers
 
-import java.io.{File, FileOutputStream}
+import lib.imaging.Projecter.createTempFile
 import java.net.URI
 
-import com.amazonaws.services.s3.model.S3Object
 import com.gu.mediaservice.lib.{DateTimeUtils, ImageIngestOperations}
 import com.gu.mediaservice.lib.argo.ArgoHelpers
 import com.gu.mediaservice.lib.argo.model.Link
 import com.gu.mediaservice.lib.auth.Authentication.Principal
 import com.gu.mediaservice.lib.auth._
-import com.gu.mediaservice.lib.aws.{S3Ops, UpdateMessage}
+import com.gu.mediaservice.lib.aws.UpdateMessage
 import com.gu.mediaservice.lib.logging.{FALLBACK, RequestLoggingContext}
-import com.gu.mediaservice.model.{Image, UploadInfo}
+import com.gu.mediaservice.model.UploadInfo
 import lib._
-import lib.imaging.MimeTypeDetection
+import lib.imaging.{Importer, MimeTypeDetection, Projecter}
 import lib.storage.ImageLoaderStore
-import model.{ImageUploadOps, ImageUploadProjector, S3FileExtractedMetadata, UploadRequest}
+import model.{ImageUploadOps, ImageUploadProjector, UploadRequest}
 import net.logstash.logback.marker.LogstashMarker
-import org.apache.tika.io.IOUtils
 import org.joda.time.DateTime
 import play.api.Logger
 import play.api.libs.json.Json
 import play.api.libs.ws.WSClient
 import play.api.mvc._
 
-import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext, Future}
 import scala.util.control.NonFatal
 import scala.util.{Failure, Success, Try}
 
@@ -33,6 +30,9 @@ class ImageLoaderController(auth: Authentication, downloader: Downloader, store:
                             config: ImageLoaderConfig, imageUploadOps: ImageUploadOps, imageUploadProjector: ImageUploadProjector,
                             override val controllerComponents: ControllerComponents, wSClient: WSClient)(implicit val ec: ExecutionContext)
   extends BaseController with ArgoHelpers {
+
+  private val imageProjecter = new Projecter(config)
+  private val imageImporter = new Importer(config, imageUploadOps, notifications, store)
 
   val indexResponse: Result = {
     val indexData = Map("description" -> "This is the Loader Service")
@@ -43,16 +43,9 @@ class ImageLoaderController(auth: Authentication, downloader: Downloader, store:
     respond(indexData, indexLinks)
   }
 
-  def index = auth {
-    indexResponse
-  }
+  def index: Action[AnyContent] = auth { indexResponse }
 
-  private def createTempFile(prefix: String, requestContext: RequestLoggingContext) = {
-    Logger.info(s"creating temp file in ${config.tempDir}")(requestContext.toMarker())
-    File.createTempFile(prefix, "", config.tempDir)
-  }
-
-  def loadImage(uploadedBy: Option[String], identifiers: Option[String], uploadTime: Option[String], filename: Option[String]) = {
+  def loadImage(uploadedBy: Option[String], identifiers: Option[String], uploadTime: Option[String], filename: Option[String]): Action[DigestedFile] = {
     val requestContext = RequestLoggingContext(
       initialMarkers = Map(
         "requestType" -> "load-image"
@@ -68,23 +61,17 @@ class ImageLoaderController(auth: Authentication, downloader: Downloader, store:
 
     Logger.info("loadImage request start")(requestContext.toMarker(markers))
 
-    val parsedBody = DigestBodyParser.create(createTempFile("requestBody", requestContext))
+    val parsedBody = DigestBodyParser.create(createTempFile(config, "requestBody", requestContext))
     Logger.info("body parsed")(requestContext.toMarker(markers))
 
     auth.async(parsedBody) { req =>
-      val result = loadFile(req.body, req.user, uploadedBy, identifiers, DateTimeUtils.fromValueOrNow(uploadTime), filename, requestContext)
+      val result = imageImporter.loadFile(req.body, req.user, uploadedBy, identifiers, DateTimeUtils.fromValueOrNow(uploadTime), filename.flatMap(_.trim.nonEmptyOpt), requestContext)
       Logger.info("loadImage request end")(requestContext.toMarker(markers))
       result
     }
   }
 
-  private def getSrcFileDigestForProjection(s3Src: S3Object, imageId: String, requestLoggingContext: RequestLoggingContext) = {
-    val uploadedFile = createTempFile(s"projection-${imageId}", requestLoggingContext)
-    IOUtils.copy(s3Src.getObjectContent, new FileOutputStream(uploadedFile))
-    DigestedFile(uploadedFile, imageId)
-  }
-
-  def projectImageBy(imageId: String) = {
+  def projectImageBy(imageId: String): Action[AnyContent] = {
     val requestContext = RequestLoggingContext(
       initialMarkers = Map(
         "imageId" -> imageId,
@@ -93,27 +80,31 @@ class ImageLoaderController(auth: Authentication, downloader: Downloader, store:
     )
 
     auth { _ =>
-      projectS3ImageById(imageId, requestContext) match {
+      imageProjecter.projectS3ImageById(imageUploadProjector, imageId, requestContext) match {
         case Success(maybeImage) =>
           maybeImage match {
-            case Some(img) => {
+            case Some(img) =>
               Logger.info("image found")(requestContext.toMarker())
               Ok(Json.toJson(img)).as(ArgoMediaType)
-            }
             case None =>
               val s3Path = "s3://" + config.imageBucket + "/" + ImageIngestOperations.fileKeyFromId(imageId)
               Logger.info("image not found")(requestContext.toMarker())
               respondError(NotFound, "image-not-found", s"Could not find image: $imageId in s3 at $s3Path")
           }
-        case Failure(error) => {
+        case Failure(error) =>
           Logger.error(s"image projection failed", error)(requestContext.toMarker())
           respondError(InternalServerError, "image-projection-failed", error.getMessage)
-        }
       }
     }
   }
 
-  def importImage(uri: String, uploadedBy: Option[String], identifiers: Option[String], uploadTime: Option[String], filename: Option[String]) = {
+  def importImage(
+                   uri: String,
+                   uploadedBy: Option[String],
+                   identifiers: Option[String],
+                   uploadTime: Option[String],
+                   filename: Option[String]
+                 ): Action[AnyContent] = {
     auth.async { request =>
       val requestContext = RequestLoggingContext(
         initialMarkers = Map(
@@ -127,10 +118,10 @@ class ImageLoaderController(auth: Authentication, downloader: Downloader, store:
         "key-name" -> apiKey.identity
       )))
       Try(URI.create(uri)) map { validUri =>
-        val tmpFile = createTempFile("download", requestContext)
+        val tmpFile = createTempFile(config, "download", requestContext)
 
         val result = downloader.download(validUri, tmpFile).flatMap { digestedFile =>
-          loadFile(digestedFile, request.user, uploadedBy, identifiers, DateTimeUtils.fromValueOrNow(uploadTime), filename, requestContext)
+          imageImporter.loadFile(digestedFile, request.user, uploadedBy, identifiers, DateTimeUtils.fromValueOrNow(uploadTime), filename.flatMap(_.trim.nonEmptyOpt), requestContext)
         } recover {
           case NonFatal(e) =>
             Logger.error(s"Unable to download image $uri", e)
@@ -154,96 +145,6 @@ class ImageLoaderController(auth: Authentication, downloader: Downloader, store:
     }
   }
 
-  private def projectS3ImageById(imageId: String, requestLoggingContext: RequestLoggingContext): Try[Option[Image]] = {
-    Logger.info(s"projecting image: $imageId")(requestLoggingContext.toMarker())
-
-    import ImageIngestOperations.fileKeyFromId
-    val s3Key = fileKeyFromId(imageId)
-    val s3 = S3Ops.buildS3Client(config)
-
-    if (!s3.doesObjectExist(config.imageBucket, s3Key)) return Try(None)
-
-    Logger.info(s"object exists, getting s3 object at s3://${config.imageBucket}/$s3Key to perform Image projection")(requestLoggingContext.toMarker())
-
-    val s3Source = s3.getObject(config.imageBucket, s3Key)
-    val digestedFile = getSrcFileDigestForProjection(s3Source, imageId, requestLoggingContext)
-    val extractedS3Meta = S3FileExtractedMetadata(s3Source.getObjectMetadata)
-
-    Try {
-      val finalImageFuture = imageUploadProjector.projectImage(digestedFile, extractedS3Meta, requestLoggingContext)
-      val finalImage = Await.result(finalImageFuture, Duration.Inf)
-      Some(finalImage)
-    }
-  }
-
-  private def loadFile(digestedFile: DigestedFile, user: Principal,
-               uploadedBy: Option[String], identifiers: Option[String],
-               uploadTime: DateTime, filename: Option[String], requestLoggingContext: RequestLoggingContext): Future[Result] = {
-    val DigestedFile(tempFile_, id_) = digestedFile
-
-    val uploadedBy_ = uploadedBy match {
-      case Some(by) => by
-      case None => Authentication.getIdentity(user)
-    }
-
-    // TODO: should error if the JSON parsing failed
-    val identifiers_ = identifiers.map(Json.parse(_).as[Map[String, String]]) getOrElse Map()
-
-    val uploadInfo_ = UploadInfo(filename.flatMap(_.trim.nonEmptyOpt))
-
-    Logger.info("Detecting mimetype")(requestLoggingContext.toMarker())
-    // Abort early if unsupported mime-type
-    val mimeType_ = MimeTypeDetection.guessMimeType(tempFile_)
-    Logger.info(s"Detected mimetype as ${mimeType_.getOrElse(FALLBACK)}")(requestLoggingContext.toMarker())
-
-    val uploadRequest = UploadRequest(
-      requestId = requestLoggingContext.requestId,
-      imageId = id_,
-      tempFile = tempFile_,
-      mimeType = mimeType_,
-      uploadTime = uploadTime,
-      uploadedBy = uploadedBy_,
-      identifiers = identifiers_,
-      uploadInfo = uploadInfo_
-    )
-
-    val supportedMimeType = config.supportedMimeTypes.exists(mimeType_.contains(_))
-
-    if (supportedMimeType) {
-      storeFile(uploadRequest, requestLoggingContext)
-    } else {
-      Future {
-        FailureResponse.unsupportedMimeType(uploadRequest, config.supportedMimeTypes)
-      }
-    }
-  }
-
-  def storeFile(uploadRequest: UploadRequest, requestLoggingContext: RequestLoggingContext): Future[Result] = {
-    val completeMarkers: LogstashMarker = requestLoggingContext.toMarker().and(uploadRequest.toLogMarker)
-
-    Logger.info("Storing file")(completeMarkers)
-    val result = for {
-      imageUpload <- imageUploadOps.fromUploadRequest(uploadRequest, requestLoggingContext)
-      image = imageUpload.image
-    } yield {
-      val updateMessage = UpdateMessage(subject = "image", image = Some(image))
-      notifications.publish(updateMessage)
-
-      // TODO: centralise where all these URLs are constructed
-      Accepted(Json.obj("uri" -> s"${config.apiUri}/images/${uploadRequest.imageId}")).as(ArgoMediaType)
-    }
-
-    result recover {
-      case e =>
-        Logger.warn(s"Failed to store file: ${e.getMessage}.", e)(completeMarkers)
-
-        store.deleteOriginal(uploadRequest.imageId).onComplete {
-          case Failure(err) => Logger.error(s"Failed to delete image for ${uploadRequest.imageId}: $err")(completeMarkers)
-          case _ =>
-        }
-        respondError(BadRequest, "upload-error", e.getMessage)
-    }
-  }
 
   // Find this a better home if used more widely
   implicit class NonEmpty(s: String) {
